Moved Tables.cs to a new namespace
